Corrected a bug in functions.php.
Added a new variable in configure.php that manages the time stamp checking variance for the reminder script.

Created checkReminder.php which is meant to be run from the command line or as a cron job.  It is the script that checks for and sends out appropriate reminders to subscribers.  It currently only supports subscribers by event type.